+++ test/integration/test_call_variant_peptides.py
@@ -610,3 +610,14 @@ class TestCallVariantPeptides(TestCaseIntegration):
         expected = self.data_dir/'fuzz/08/brute_force.txt'
         reference = self.data_dir/'downsampled_reference/ENST00000314675.11'
         self.default_test_case(gvf, reference, expected)
+
+    def test_call_variant_peptide_case37(self):
+        """ Noncoding TX with stop lost mutation. This ensures that the
+        start gain and stop lost mutations before the novel start site are not
+        retained. #519 """
+        gvf = [
+            self.data_dir/'fuzz/09/fake_variants.gvf'
+        ]
+        expected = self.data_dir/'fuzz/09/brute_force.txt'
+        reference = self.data_dir/'downsampled_reference/ENST00000452737.5'
+        self.default_test_case(gvf, reference, expected)